update to CosmWasm 1.1 and reorganize workspace
