--- contracts/cw-ibc-loan-receiver/src/state.rs
@@ -1,4 +1,0 @@
-use cosmwasm_std::IbcMsg;
-use cw_storage_plus::Item;
-
-pub const MSG: Item<IbcMsg> = Item::new("msg");
