--- contracts/cw-simple-loan-receiver/.cargo/config
@@ -1,4 +1,0 @@
-[alias]
-wasm = "build --release --target wasm32-unknown-unknown"
-unit-test = "test --lib"
-schema = "run --example schema"
