--- contracts/cw-simple-loan-receiver/schema/instantiate_msg.json
@@ -1,14 +1,0 @@
-{
-  "$schema": "http://json-schema.org/draft-07/schema#",
-  "title": "InstantiateMsg",
-  "type": "object",
-  "required": [
-    "count"
-  ],
-  "properties": {
-    "count": {
-      "type": "integer",
-      "format": "int32"
-    }
-  }
-}
