--- contracts/cw-simple-loan-receiver/schema/query_msg.json
@@ -1,18 +1,0 @@
-{
-  "$schema": "http://json-schema.org/draft-07/schema#",
-  "title": "QueryMsg",
-  "anyOf": [
-    {
-      "type": "object",
-      "required": [
-        "get_count"
-      ],
-      "properties": {
-        "get_count": {
-          "type": "object"
-        }
-      },
-      "additionalProperties": false
-    }
-  ]
-}
