--- contracts/cw-simple-loan-receiver/src/lib.rs
@@ -1,6 +1,0 @@
-pub mod contract;
-mod error;
-pub mod msg;
-pub mod state;
-
-pub use crate::error::ContractError;
