+++ debug/cw-simple-loan-receiver/.cargo/config
