+++ debug/cw-simple-loan-receiver/schema/count_response.json
