+++ debug/cw-simple-loan-receiver/schema/execute_msg.json
