+++ debug/cw-simple-loan-receiver/schema/instantiate_msg.json
