+++ debug/cw-simple-loan-receiver/schema/query_msg.json
