+++ debug/cw-simple-loan-receiver/schema/state.json
