+++ debug/cw-simple-loan-receiver/src/contract.rs
