+++ debug/cw-simple-loan-receiver/src/lib.rs
