+++ debug/cw-simple-loan-receiver/src/state.rs
